+++ lib/src/components/checkbox.dart
@@ -32,6 +32,7 @@ class ShadCheckbox extends StatefulWidget {
     this.padding,
     this.direction,
     this.crossAxisAlignment,
+    this.checkboxPadding,
   });
 
   /// {@template ShadCheckbox.value}
@@ -102,9 +103,16 @@ class ShadCheckbox extends StatefulWidget {
 
   /// {@template ShadCheckbox.padding}
   /// The padding between the checkbox and its label/sublabel.
-  /// Defaults to EdgeInsets.only(left: 8) if not specified.
+  /// Defaults to `EdgeInsets.only(left: 8)` if not specified.
   /// {@endtemplate}
-  final EdgeInsets? padding;
+  final EdgeInsetsGeometry? padding;
+
+  /// {@template ShadCheckbox.checkboxPadding}
+  /// The padding around the checkbox field only.
+  ///
+  /// Defaults to `EdgeInsets.only(top: 1)` if not specified.
+  /// {@endtemplate}
+  final EdgeInsetsGeometry? checkboxPadding;
 
   /// {@template ShadCheckbox.direction}
   /// The text direction for arranging the checkbox and label/sublabel.
@@ -177,6 +185,10 @@ class _ShadCheckboxState extends State<ShadCheckbox> {
         theme.checkboxTheme.padding ??
         const EdgeInsets.only(left: 8);
 
+    final effectiveCheckboxPadding = widget.checkboxPadding ??
+        theme.checkboxTheme.checkboxPadding ??
+        const EdgeInsets.only(top: 1);
+
     final checkbox = Semantics(
       checked: widget.value,
       child: ShadDisabled(
@@ -236,7 +248,10 @@ class _ShadCheckboxState extends State<ShadCheckbox> {
           mainAxisSize: MainAxisSize.min,
           crossAxisAlignment: effectiveCrossAxisAlignment,
           children: [
-            checkbox,
+            Padding(
+              padding: effectiveCheckboxPadding,
+              child: checkbox,
+            ),
             if (widget.label != null || widget.sublabel != null)
               Flexible(
                 child: Padding(